refactor(Name): 클래스 재사용을 위한 리팩터링
- record 클래스에서 변경
- 클래스 의미 변화로 인한 패키지 이동
