+++ src/main/java/domain/player/Player.java
@@ -1,4 +1,6 @@
 package domain.player;
 
+import domain.Name;
+
 public record Player(Name name) {
 }
