+++ src/test/java/domain/player/NamesTest.java
@@ -1,5 +1,6 @@
 package domain.player;
 
+import domain.Names;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
