+++ src/test/java/domain/player/PlayerTest.java
@@ -1,5 +1,6 @@
 package domain.player;
 
+import domain.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
