+++ src/test/java/domain/player/PlayersTest.java
@@ -1,5 +1,7 @@
 package domain.player;
 
+import domain.Name;
+import domain.Names;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
